Clean up gitignores
It is preferred to list only project files to ignore in the repository's
own `.gitignore` and list stuff that should **always** be ignored in the
user's global ignore file.

> So that's the general idea: your project (repository) .gitignore
  should list the names or name-patterns of files that will be found in
  work-trees when working with your project, but that should not be
  committed to the project. In other words, it's not OS-specific, it's
  project-specific. Other file name patterns—OS-specific,
  editor-specific, IDE-specific, and so on—can go in other ignore files,
  and hence need not be listed in the project's .gitignore files. It
  does not necessarily hurt to list them in the project's files, but if
  each person is sensible about things, it doesn't help either.

Other files that should not always be ignored and are not supposed to be
encountered in this project (such as yarn files, bower files, etc.)
simply should not be listed at all.

See: https://git-scm.com/docs/gitignore#_configuration
See: https://stackoverflow.com/questions/61982834/what-should-contain-gitignore-file-when-is-a-public-repository